fix(deploy): mount automatic GitHub authorization before connection
